Update redux store
* Move simulation selectors in simulation reducer
* Update simulation reducer
* TODO: fetch preprovisionedhub in simulation API
